再次提交一堆东西

+++ src/main/java/cc/niushuai/graduate/entity/EduCourseSubject.java
@@ -1,17 +1,12 @@
 package cc.niushuai.graduate.entity;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.util.Date;
-
 
 
 /**
- * 
- * 
  * @author niushuai
  * @email niushuai951101@gmail.com
  * @date 2019-05-07 22:08:20
@@ -19,32 +14,25 @@ import java.util.Date;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class EduCourseSubject implements Serializable {
-	private static final long serialVersionUID = 1L;
-	
+    private static final long serialVersionUID = 1L;
 
 
-
-	/**
-	 *
-	 * 
-	 */
-		private Integer id;
+    /**
+     *
+     */
+    private Long id;
 
 
-
-	/**
-	 *
-	 * 课程id
-	 */
-		private Integer courseId;
+    /**
+     * 课程id
+     */
+    private Long courseId;
 
 
-
-	/**
-	 *
-	 * 分类id
-	 */
-		private Integer subjectId;
+    /**
+     * 分类id
+     */
+    private Long subjectId;
 
 
 }
