+++ src/main/java/cc/niushuai/graduate/entity/EduCourseTeacher.java
@@ -1,17 +1,14 @@
 package cc.niushuai.graduate.entity;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.util.Date;
-
 
 
 /**
  * 课程讲师关联
- * 
+ *
  * @author niushuai
  * @email niushuai951101@gmail.com
  * @date 2019-05-07 22:06:27
@@ -19,24 +16,19 @@ import java.util.Date;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class EduCourseTeacher implements Serializable {
-	private static final long serialVersionUID = 1L;
-	
+    private static final long serialVersionUID = 1L;
 
 
-
-	/**
-	 *
-	 * 课程id
-	 */
-		private Integer courseId;
+    /**
+     * 课程id
+     */
+    private Long courseId;
 
 
-
-	/**
-	 *
-	 * 讲师id
-	 */
-		private Integer teacherId;
+    /**
+     * 讲师id
+     */
+    private Long teacherId;
 
 
 }
